+++ lib/services/local/saml_idp_service_provider_test.go
@@ -34,6 +34,7 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/backend/memory"
+	"github.com/gravitational/teleport/lib/services"
 )
 
 // TestSAMLIdPServiceProviderCRUD tests backend operations with SAML IdP service provider resources.
@@ -206,19 +207,6 @@ const testEntityDescriptor = `<?xml version="1.0" encoding="UTF-8"?>
 </md:EntityDescriptor>
 `
 
-func newSAMLSPMetadata(entityID, acsURL string) string {
-	return fmt.Sprintf(samlSPMetadata, entityID, acsURL)
-}
-
-// samlSPMetadata mimics metadata generated by saml.ServiceProvider.Metadata()
-const samlSPMetadata = `<EntityDescriptor xmlns="urn:oasis:names:tc:SAML:2.0:metadata" validUntil="2023-12-09T23:43:58.16Z" entityID="%s">
-<SPSSODescriptor xmlns="urn:oasis:names:tc:SAML:2.0:metadata" validUntil="2023-12-09T23:43:58.16Z" protocolSupportEnumeration="urn:oasis:names:tc:SAML:2.0:protocol" AuthnRequestsSigned="false" WantAssertionsSigned="true">
-  <NameIDFormat>urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified</NameIDFormat>
-  <AssertionConsumerService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST" Location="%s" index="1"></AssertionConsumerService>
-</SPSSODescriptor>
-</EntityDescriptor>
-`
-
 func TestCreateSAMLIdPServiceProvider_fetchOrGenerateEntityDescriptor(t *testing.T) {
 	ctx := context.Background()
 	backend, err := memory.New(memory.Config{
@@ -301,7 +289,7 @@ func TestCreateSAMLIdPServiceProvider_fetchOrGenerateEntityDescriptor(t *testing
 	sp2FromBackend, err := service2.GetSAMLIdPServiceProvider(ctx, sp2.GetName())
 	require.NoError(t, err)
 
-	metadataTemplate := newSAMLSPMetadata(notFoundURL, testSPServer.URL)
+	metadataTemplate := services.NewSAMLTestSPMetadata(notFoundURL, testSPServer.URL)
 
 	expected, err := samlsp.ParseMetadata([]byte(metadataTemplate))
 	require.NoError(t, err)
@@ -327,7 +315,7 @@ func TestCreateSAMLIdPServiceProvider_fetchAndSetEntityDescriptor(t *testing.T) 
 			fmt.Fprintln(w, "test")
 		default:
 			location := fmt.Sprintf("https://%s", r.Host)
-			metadata := newSAMLSPMetadata(location, location)
+			metadata := services.NewSAMLTestSPMetadata(location, location)
 			w.WriteHeader(http.StatusOK)
 			fmt.Fprintln(w, metadata)
 		}
@@ -598,176 +586,4 @@ func TestCreateSAMLIdPServiceProvider_GetTeleportSPSSODescriptor(t *testing.T) {
 	})
 	index, _ := GetTeleportSPSSODescriptor(ed.SPSSODescriptors)
 	require.Equal(t, 3, index)
-}
-
-func TestCreateSAMLIdPServiceProviderInvalidInputs(t *testing.T) {
-	ctx := context.Background()
-
-	backend, err := memory.New(memory.Config{
-		Context: ctx,
-		Clock:   clockwork.NewFakeClock(),
-	})
-	require.NoError(t, err)
-
-	service, err := NewSAMLIdPServiceProviderService(backend)
-	require.NoError(t, err)
-
-	tests := []struct {
-		name             string
-		entityDescriptor string
-		entityID         string
-		acsURL           string
-		errAssertion     require.ErrorAssertionFunc
-	}{
-		{
-			name:     "missing url scheme in acs input",
-			entityID: "sp",
-			acsURL:   "sp",
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid scheme")
-			},
-		},
-		{
-			name:             "missing url scheme for acs in ed",
-			entityDescriptor: newSAMLSPMetadata("sp", "sp"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid url scheme")
-			},
-		},
-		{
-			name:     "http url scheme in acs",
-			entityID: "sp",
-			acsURL:   "http://sp",
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid scheme")
-			},
-		},
-		{
-			name:             "http url scheme for acs in ed",
-			entityDescriptor: newSAMLSPMetadata("sp", "http://sp"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "unsupported ACS bindings")
-			},
-		},
-		{
-			name:     "unsupported scheme in acs",
-			entityID: "sp",
-			acsURL:   "gopher://sp",
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid scheme")
-			},
-		},
-		{
-			name:             "unsupported scheme for acs in ed",
-			entityDescriptor: newSAMLSPMetadata("sp", "gopher://sp"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid url scheme")
-			},
-		},
-		{
-			name:     "invalid character in acs",
-			entityID: "sp",
-			acsURL:   "https://sp>",
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "unsupported character")
-			},
-		},
-		{
-			name:             "invalid character in acs in ed",
-			entityDescriptor: newSAMLSPMetadata("sp", "https://sp>"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "unsupported ACS bindings")
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			sp, err := types.NewSAMLIdPServiceProvider(types.Metadata{
-				Name: "test",
-			}, types.SAMLIdPServiceProviderSpecV1{
-				EntityDescriptor: test.entityDescriptor,
-				EntityID:         test.entityID,
-				ACSURL:           test.acsURL,
-			})
-			require.NoError(t, err)
-
-			err = service.CreateSAMLIdPServiceProvider(ctx, sp)
-			test.errAssertion(t, err)
-		})
-	}
-}
-
-func TestUpdateSAMLIdPServiceProviderInvalidInputs(t *testing.T) {
-	ctx := context.Background()
-
-	backend, err := memory.New(memory.Config{
-		Context: ctx,
-		Clock:   clockwork.NewFakeClock(),
-	})
-	require.NoError(t, err)
-
-	service, err := NewSAMLIdPServiceProviderService(backend)
-	require.NoError(t, err)
-
-	sp, err := types.NewSAMLIdPServiceProvider(types.Metadata{
-		Name: "sp",
-	}, types.SAMLIdPServiceProviderSpecV1{
-		EntityDescriptor: newSAMLSPMetadata("https://sp", "https://sp"),
-	})
-	require.NoError(t, err)
-
-	err = service.CreateSAMLIdPServiceProvider(ctx, sp)
-	require.NoError(t, err)
-
-	tests := []struct {
-		name             string
-		entityDescriptor string
-		entityID         string
-		acsURL           string
-		errAssertion     require.ErrorAssertionFunc
-	}{
-		{
-			name:             "missing url scheme for acs in ed",
-			entityDescriptor: newSAMLSPMetadata("https://sp", "sp"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid url scheme")
-			},
-		},
-		{
-			name:             "http url scheme for acs in ed",
-			entityDescriptor: newSAMLSPMetadata("https://sp", "http://sp"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "unsupported ACS bindings")
-			},
-		},
-		{
-			name:             "unsupported scheme for acs in ed",
-			entityDescriptor: newSAMLSPMetadata("https://sp", "gopher://sp"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "invalid url scheme")
-			},
-		},
-		{
-			name:             "invalid character in acs in ed",
-			entityDescriptor: newSAMLSPMetadata("https://sp", "https://sp>"),
-			errAssertion: func(t require.TestingT, err error, i ...interface{}) {
-				require.ErrorContains(t, err, "unsupported ACS bindings")
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			sp, err := types.NewSAMLIdPServiceProvider(types.Metadata{
-				Name: "sp",
-			}, types.SAMLIdPServiceProviderSpecV1{
-				EntityDescriptor: test.entityDescriptor,
-			})
-			require.NoError(t, err)
-
-			err = service.UpdateSAMLIdPServiceProvider(ctx, sp)
-			test.errAssertion(t, err)
-		})
-	}
 }
